+++ ormar/queryset/utils.py
@@ -18,7 +18,7 @@ if TYPE_CHECKING:  # pragma no cover
 
 
 def check_node_not_dict_or_not_last_node(
-    part: str, parts: List, current_level: Any
+    part: str, is_last: bool, current_level: Any
 ) -> bool:
     """
     Checks if given name is not present in the current level of the structure.
@@ -36,7 +36,7 @@ def check_node_not_dict_or_not_last_node(
     :return: result of the check
     :rtype: bool
     """
-    return (part not in current_level and part != parts[-1]) or (
+    return (part not in current_level and not is_last) or (
         part in current_level and not isinstance(current_level[part], dict)
     )
 
@@ -71,9 +71,10 @@ def translate_list_to_dict(  # noqa: CCR001
             else:
                 def_val = "asc"
 
-        for part in parts:
+        for ind, part in enumerate(parts):
+            is_last = ind == len(parts) - 1
             if check_node_not_dict_or_not_last_node(
-                part=part, parts=parts, current_level=current_level
+                part=part, is_last=is_last, current_level=current_level
             ):
                 current_level[part] = dict()
             elif part not in current_level:
@@ -124,6 +125,49 @@ def update(current_dict: Any, updating_dict: Any) -> Dict:  # noqa: CCR001
             current_dict[key] = current_dict.get(key).union(value)
         else:
             current_dict[key] = value
+    return current_dict
+
+
+def subtract_dict(current_dict: Any, updating_dict: Any) -> Dict:  # noqa: CCR001
+    """
+    Update one dict with another but with regard for nested keys.
+
+    That way nested sets are unionised, dicts updated and
+    only other values are overwritten.
+
+    :param current_dict: dict to update
+    :type current_dict: Dict[str, ellipsis]
+    :param updating_dict: dict with values to update
+    :type updating_dict: Dict
+    :return: combination of both dicts
+    :rtype: Dict
+    """
+    for key, value in updating_dict.items():
+        old_key = current_dict.get(key, {})
+        new_value: Optional[Union[Dict, Set]] = None
+        if not old_key:
+            continue
+        if isinstance(value, set) and isinstance(old_key, set):
+            new_value = old_key.difference(value)
+        elif isinstance(value, (set, collections.abc.Mapping)) and isinstance(
+            old_key, (set, collections.abc.Mapping)
+        ):
+            value = (
+                convert_set_to_required_dict(value)
+                if not isinstance(value, collections.abc.Mapping)
+                else value
+            )
+            old_key = (
+                convert_set_to_required_dict(old_key)
+                if not isinstance(old_key, collections.abc.Mapping)
+                else old_key
+            )
+            new_value = subtract_dict(old_key, value)
+
+        if new_value:
+            current_dict[key] = new_value
+        else:
+            current_dict.pop(key, None)
     return current_dict